Made Radi and ComponentClass one
