+++ src/consts/GLOBALS.js
@@ -3,10 +3,8 @@ const GLOBALS = {
   REGISTERED: {},
   MIX: {},
   FROZEN_STATE: false,
-  RL: "(".charCodeAt(0),
-  RR: ")".charCodeAt(0),
   IDS: 0,
-  VERSION: "0.1.8",
+  VERSION: '0.1.8',
   ACTIVE_COMPONENTS: [],
   HTML_CACHE: {},
   R_KEYS: 0
