+++ src/mount.js
@@ -1,7 +1,7 @@
 const mount = (component, id) => {
   const container = typeof id === 'string' ? document.getElementById(id) : id;
   const rendered =
-    component instanceof Component ? component.__radi().$render() : component;
+    component instanceof Component ? component.$render() : component;
   container.appendChild(rendered);
   return rendered;
 };
